backend/requirements.txt
Django==5.1.6
psycopg2-binary==2.9.10
dj-database-url==2.2.0
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.4.0
django-filter==24.3
python-dotenv==1.0.1
